WIP on develop feat(map): Add disable cluster layer functionality for related pois
This commit adds the `[wmMapReletedPoisDisableClusterLayer]` attribute to the map page HTML template. The attribute is bound to a condition that disables the cluster layer for related pois when certain conditions are met. This feature enhances the user experience by providing more control over the display of related pois on the map.

WIP on develop