chore: Add WmITrackToFeaturePipe to pipe module oc: 4071
This commit adds the `WmITrackToFeaturePipe` to the `PipeModule` in order to transform an `ITrack` object into a GeoJSON `Feature`. The pipe takes an `ITrack` object as input and returns a GeoJSON `Feature` with the necessary properties and coordinates. This pipe is used in the map page template to convert the track data before displaying it on the map.