chore(draw-track): add createdAt and updatedAt properties to track oc: 4453 (#28)
* chore(draw-track): add createdAt and updatedAt properties to track

This commit adds the `createdAt` and `updatedAt` properties to the track object in the `DrawTrackComponent`. These properties are set to the current date and time when a new track is created or updated.

Updated submodule src/app/shared/wm-core

Updated submodule src/app/shared/wm-types

* chore: Update map.page.ts

This commit updates the map.page.ts file. It adds a return statement of null in two places, which improves the code logic and functionality.

Updated submodule src/app/shared/wm-core

* Updated submodule src/app/shared/map-core

* Updated submodule src/app/shared/wm-core