feat(map): Add support for setting current related POI
This commit adds support for setting the current related POI in the map page. The `setCurrentRelatedPoi` method now accepts a parameter of type `number`, `WmFeature<Point>`, or `null`. If the parameter is `null`, the method returns early. If it's a number, the URL and directive are updated accordingly. If it's an object with properties, the ID property is used to update the URL and directive.

This change improves the functionality of the map page by allowing users to set the current related POI more flexibly.